% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{hello_admiral}
\alias{hello_admiral}
\title{Derive Extension Example}
\usage{
hello_admiral(hw = TRUE)
}
\arguments{
\item{hw}{Boolean determining whether a hello message is returned.

\describe{
\item{Permitted values}{\code{TRUE}, \code{FALSE}}
\item{Default value}{\code{TRUE}}
}}
}
\value{
A message
}
\description{
Says hello admiral
}
\details{
In the roxygen documentation you will find tags for family and keywords.
This is to create organized sections for the Reference tab on the pkgdown website.
You can modify the \verb{_pkgdown.yml} as necessary to create appropriate sections as necessary.
Under \code{./man/roxygen/meta.R}, you will find where to store these family/keywords.
You will also find custom permitted text under \code{./man/roxygen/rdx_meta.R} to
standardize argument text.
}
\examples{
hello_admiral()
hello_admiral(hw = FALSE)
}
\concept{der_adxx}
\keyword{der_adxx}
