% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xx}
\alias{xx}
\title{xx Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 1 columns.
}
\usage{
xx
}
\description{
An xx test dataset for xx
}
\concept{datasets}
\keyword{datasets}
